backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
beautifulsoup4
faiss-cpu
sentence-transformers
google-genai
python-dotenv

